+++ src/modules/policy/policy.auth.ts
@@ -102,6 +102,7 @@ const billStatus = 'billStatus'; // Trạng thái đơn hàng bán
 const billSpit = 'billSpit'; // Tách đơn hàng bán
 const historyBill = 'historyBill'; // Lịch sử đơn hàng bán
 const historyOrderSupplier = 'historyOrderSupplier'; // Lịch sử đơn hàng mua
+const outOfStock = 'outOfStock'; // Kiểm kho theo danh sách đơn hàng bán
 const coupon = 'coupon'; // Tách đơn hàng bán
 
 const reportOverviewB2B = 'reportOverviewB2B';
@@ -218,6 +219,7 @@ const RESOURCES = [
   billSpit,
   historyBill,
   historyOrderSupplier,
+  outOfStock,
   coupon,
   reportOverviewB2B,
   reportOverviewB2C,
@@ -385,9 +387,13 @@ const RESOURCE = (): ResourceType => {
   const GROUP_SETTING: string[] = [
     warehouseLink,
     ];
+  const WAREHOUSE: string[] = [
+    outOfStock
+  ];
   const REPORT_OVERVIEW: string[] = [
     reportOverview,
   ];
+
   const REPORT_OVERVIEW_B2B : string[] = [
     reportOverviewB2B
   ];
@@ -406,8 +412,7 @@ const RESOURCE = (): ResourceType => {
   const REPORT_REVENUE_TEAM_PARTNER : string [] = [
     reportRevenueTeamPartner
   ];
-
-  return {
+return {
     GROUP_USER,
     GROUP_EMPLOYEE,
     GROUP_WHSETTING,
@@ -433,6 +438,7 @@ const RESOURCE = (): ResourceType => {
     VOUCHER_SALARY_PARTNER,
     VOUCHER_BILL_PARTNER,
     GROUP_SETTING,
+    WAREHOUSE,
     REPORT_OVERVIEW,
     REPORT_OVERVIEW_B2B,
     REPORT_OVERVIEW_B2C,