+++ src/modules/sale/orderSupplier/components/createOrderSupplierScreen/SaleScreen.tsx
@@ -113,7 +113,7 @@ try {
         pair : 0
       }}
     >
-      <Row gutter={16}>
+      <Row gutter={8}>
         <Col span={16}>
           <ProductSelectedTable />
         </Col>