+++ src/modules/sale/orderSupplier/screens/UpdateOrderSupplier.tsx
@@ -93,6 +93,7 @@ export default function UpdateBill(props: propsType): React.JSX.Element {
   const [isSubmitCreate, onCreateOrderInWarehouse] = useCreateOrderInWarehouse(() => {
     dispatch(orderSupplierActions.resetActionInById());
   });
+  console.log(isSubmitCreate,'isSubmitCreate')
   const onCloseCancel = useCallback(() => {
     setOpenCancel(false);
     setCancelNote("");